cpp refactor example
