+++ c-abi-libp2p/examples/cpp/cli_args.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "abi_bindings.h"
+
+namespace ping_example {
+
+Arguments parseArgs(int argc, char** argv);
+
+} // namespace ping_example
