+++ c-abi-libp2p/examples/cpp/ping_app.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "abi_bindings.h"
+
+namespace ping_example {
+
+int runPingApp(const CabiRustLibp2p& abi, const Arguments& args);
+bool loadAbi(const class DynamicLibrary& library, CabiRustLibp2p& abi);
+
+} // namespace ping_example
